requirements.txt
asttokens==3.0.1
certifi==2026.2.25
charset-normalizer==3.4.7
comm==0.2.3
contourpy==1.3.3
cuda-bindings==12.9.4
cuda-pathfinder==1.2.2
cuda-toolkit==12.6.3
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
executing==2.2.1
filelock==3.25.2
fonttools==4.62.1
fsspec==2026.2.0
idna==3.11
ipykernel==7.2.0
ipython==9.11.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.5.0
lazy-loader==0.5
lightning-utilities==0.15.3
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mkl_fft==2.1.1
mkl_random==1.3.0
mkl-service==2.5.2
mne==1.12.0
mpmath==1.3.0
nest-asyncio==1.5.1
networkx==3.6.1
numpy==2.4.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.85
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.28.9
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.6.77
packaging==26.0
pandas==3.0.2
parso==0.8.5
pexpect==4.9.0
pillow==12.2.0
pip==26.0.1
platformdirs==4.9.4
pooch==1.9.0
prompt_toolkit==3.0.52
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.33.1
scikit-learn==1.8.0
scipy==1.17.1
setuptools==70.2.0
six==1.17.0
stack_data==0.6.3
sympy==1.14.0
threadpoolctl==3.5.0
torch==2.11.0+cu126
torchmetrics==1.9.0
torchvision==0.26.0+cu126
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
triton==3.6.0
typing_extensions==4.15.0
urllib3==2.6.3
wcwidth==0.2.14
